requirements.txt
# ProtoForge - Python Dependencies
# Install: pip install -r requirements.txt

# Core
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.1.0
requests>=2.31.0

# LangChain & AI
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=1.0.0

# Data & Validation
pydantic>=2.0.0
jsonschema>=4.23.0

# Web
aiohttp>=3.9.0
httpx>=0.27.0

# Utils
markdown>=3.5.0
pyyaml>=6.0

# Server
uvicorn>=0.30.0
